Fixed cleanup handler to through errors appropriately
--HG--
branch : trunk

+++ fileprocess/fileprocess/actions/cleanup.py
@@ -18,7 +18,5 @@ class Cleanup(BaseAction):
             file
         )
 
-        try:
+        if os.path.exists(file['fname']):
             os.remove(file['fname'])
-        except:
-            pass
